+++ lib/fullscreen_image_viewer.dart
@@ -1,28 +1,4 @@
 library fullscreen_image_viewer;
 
-import 'package:flutter/material.dart';
-import 'package:fullscreen_image_viewer/fullscreen_viewer.dart';
-
-class FullscreenImageViewer {
-  FullscreenImageViewer._();
-
-  /// [child] should be a Hero widget for transition animation
-  /// [closeWidget] is content of button at top right corner that closes viewer
-  static Future<void> open({
-    required BuildContext context,
-    required Widget child,
-    Widget? closeWidget,
-  }) {
-    return Navigator.of(context, rootNavigator: true).push(
-      PageRouteBuilder(
-        fullscreenDialog: true,
-        opaque: false,
-        pageBuilder: (context, animation, _) => FullscreenViewer(
-          openAnimation: animation,
-          closeWidget: closeWidget,
-          child: child,
-        ),
-      ),
-    );
-  }
-}
+export 'src/fullscreen_image_viewer.dart';
+export 'src/fullscreen_viewer_wrapper.dart';